Ftian sane (#3669)
* Disable show database to pass bvt.

* Bring sanity.  Dis-Assemble our code.

This is an attempt to solve #3569.   I don't know if
I fixed it, but, I have a machine that can repro the
bug in every 7 or 8 runs.   After removing assembly code
it has run 50 runs without a problem.

The assembly code is not maintainable anyway.  I cannot
debug it and I don't trust the toolchain that generated
it.  Performance wise, I have benchmarked that CGO overhead
is nothing when we set vector size to 4K or 8K and above.
Bet improvement of GCC/Clang will beat this assembly in
the long run.

Co-authored-by: bRong Njam <longran1989@gmail.com>