cmd: resolve gossip seed addresses (#4421)
When dns names are used in the GossipSeedAddresses field of logservice.Config, mo-service will try to resolve them and pass them to the logservice as a slice of strings in ip:port format. Only IPv4 addresses are kept and ErrInvalidConfig is returned if any dns name is mapped to more than one IPv4 address.

Approved by: @reusee