[bug] gossip seed num is not correct in cluster test.
gossipSeedNum() returns a wrong value and causes "shard not found"
error and cluster cannot start.

When the number of logservice is 3, the gossip seed num should
be 3 too, but not 1. So modify the gossipSeedNum() according to
its function comment. Whatever the number of logservice is, it
can return a proper value.