refactor: solve data race bug in NewReader in another way (#6556)
refactor for PR 6462 which to solve data race bug in NewReader.

the bug root cause:
1、we will make multi preScopes in compile.compileMinusAndIntersect/compileUnion/compileGroup for performance
2、multi compile.preScopes will make multi scope.ParallelRun call.  then we will get multi Reader which share one relation(another name is table) in one transaction.
3  in blockMergeReader.Read we will get the share relation's attribute (table.meta.modifedBlocks[i].deletes) to filter rows.
but while filter rows, we do a wrong action, we modify the attribute.  because multi readers exists and do Read in the same time,  data race happed!

after discuss with @nnsgmsone .  we think multi reader in one txn is necessary for performance. but reader modify share data is not allowed.  so i change the code for step 3.

Approved by: @nnsgmsone, @fengttt