add runtime and init runtime in main (#6800)
Add process-level runtime support. And init the runtime in main.

Runtime contains the runtime environment for a MO service. Each CN/DN/LOG service
needs to receive a Runtime and will pass the Runtime to all components of the service.
These Runtimes may only be created in main or integration test framework.

Because most of our BVT tests and integration tests are run in a single mo-service
process, which runs multiple CN, DN and LOG services, the Runtime cannot be set as a
global variable, otherwise we would not be able to set a single Runtime for each service
and each component.

In other words, there are no global variables inside mo-service, and the scope of global
variables is in a Runtime.

Since each component holds a Runtime, all mo-service-level parameters should not appear in
the parameter list of the component's initialization function, but should be replaced by the
Runtime's global variables.

Unfortunately, we can't fully support the isolation of the service-level runtime between services
and they still share a process-level runtime, but at least there is an advantage in that all
service-level components are placed in the runtime, so that the configuration and components can be
easily retrieved from the runtime without having to pass them as parameters, so that when adding
service-level components in the future, there is no need to modify the parameter list.

Approved by: @fengttt, @reusee, @yingfeng