refactor the function evaluate framework. (#7339)
the old function evaluate framework requires developers allocate the use the memory carefully. it's too difficult and too hard to debug when bug occurs.

New framework can let developers needn't care about if the function parameter is scalar or not, and needn't care about how to allocate the memory for vector (for much built-in functions, developers has no need to know what is vector and how to use vector.).
it brings new way to implement a built-in function, we support GetValue and Append method to get value of function parameter and set value for function result even though it's Null.

and I rewrite the built-in function `CAST` and `rand()` as a demo at this pr.

Approved by: @nnsgmsone, @aressu1985, @ouyuanning, @aunjgr