logservice: support domain name config for gossip-address.
Currenty, the config gossip-address for logservice can only be set
to an IP address, as memeberlist and dragonboat force it. But when
a logservice pod is recreated, the IP address of it is changed but
the pod's data, UUID and domain name are unchanged, and the memberlist
mark the IP address conflict and the whole cluster is unstable.

So we need to set gossip-address to a domain name as it is not changed.
Add a new config gossip-address-v2 to keep compatibility. If
gossip-address-v2 is set, use it as the gossip address, otherwise, use
the old one gossip-address.