[bug] kill query failed in multi cn env.
First, the connection id is obtained from the HAKeeper to ensure
that it remains globally unique among multiple CNs.

Second, an event mechanism is added to the proxy module, which
recognizes the SQL command by regular expressions. When it is kill
query command, it does not forward the request through the current
tunnel, but creates a new connection to the specified CN server,
executes the command, and then returns the result to the client.