[bug] proxy: support the Ctrl+C operation to kill query.
Ctrl+C operation is captured by MySQL client and send a kill
query statement to server. The connection ID is the one that client
received from server when login, which is generated by proxy. We
build a map from proxy connection ID to backend server connection
ID. The connection ID is globally unique, so we can find the correct
one at last.