Avoid redundant allocations when filtering zonemaps (#9441)
We previously allocates too many small objects when filtering zonemaps.
That's the number of nodes on the expression tree multiplied by number
of zonemaps. Actually we only need to keep a ZM and a Vector for each
node on the expression tree. They can be reused no matter how many
zonemaps we are iterating.