[refactor] proxy: some refactor on proxy for sys tenant.
1. priority of cn labels for sys tenant.
The following strategies are listed in order of priority：
(1) The CN servers which are configured as sys account.
(2) The CN servers which are configured as some labels whose key
  is not account.
(3) The CN servers which are configured as no labels.
(4) At last, if no CN servers are selected,
  (4.1) If the username is dump or root, we just select one randomly.
  (4.2) Else, no servers are selected.

2. Add retry mechanism when connect failed to cn server if
the cn server is dropped. When retry to connect to backend servers,
the failed backend servers will be exclued.