ensure txn close (#9753)
Ensure the close of an opened transaction.
If there is no opened transaction, it does nothing.
If there is an opened transaction, it rollback the transaction.


We encounter a situation:
Client (timeout 20s) sends a query the mo.
Mo run the query (more than 40s).
Client timeout and sends a tcp FIN to the mo.
The Conn does not close immediately.
The query responds to the client.(40s later)
The gotty(mo’s tcp server) finds the conn is disconnected and callback the Closed interface to clean the resources.
The mo cleans the resources but neglects to rollback the transaction.

Approved by: @qingxinhome