remove independent txn in sequence's function (#9777)
make the execution of the sequence's function in only one transaction.
the transaction can be shared among the sessions. we only use the shared transaction in the sequence currently.

in previous design, the sequence's functions (nextval,setval,curval) are ran in an independent transaction.
Base on pervious conclusion, we need them in one shared transaction.

Approved by: @ouyuanning, @fengttt, @jianwan0214, @qingxinhome, @m-schen