improve auto-increment in load scene (#9796)
For the load scenario, if the machine is good enough, there will be very many goroutines to concurrently fetch the value of the auto-increment column, which will immediately trigger the cache of the auto-increment column to be insufficient and thus go to the store to allocate a new cache, which causes the load to block all due to this allocation being slow. The idea of our optimization here is to reduce the number of allocations as much as possible, add an atomic counter, check how many concurrent requests are waiting when allocating (of course this is imprecise, but it doesn't matter), and then allocate more than one at a time.

Approved by: @sukki37