[bug] txn: drop table reported error ExpectedEOB.
Reason:
1. When the truncate table txn commit, the session which is
executing the drop table does not check the timestamp returned by
the doLock function. If the timestamp is not empty, a retry error
should be returned.
2. And if we return a retry error, the table in current workspace
should be deleted. Otherwise, the retry operation will be failed
because it uses the old table ID. And if table is deleted, a new
table will be fetched from the MO catalog.

Fix: Return a retry error if the timestamp is not emtpy and delete
the old table from current workspace.