[bug] txn: drop table should reported table not exists.
Reason:

1. For RC isolation, each statement has its own timestamp including
the first statement of the transaction. But currently, the first
statement's timestamp is the same as the transaction's. If the
first statement has some conflict, we got inconsonant.

2. Snapshot TS updating happens after build a plan, and causes the
TS is not correct.

3. there is tableMap in each transaction which is not affected by
other transactions, causing RC is broken.

Fix:
1. Make every statement has its own timestamp;
2. Move the snapshot TS updating before plan building;
3. create a table guard to update the table map in the transaction.
If another transaction drop/truncate a table, all other transactions
perceive it and will update their table map.