reduce the number of workers to execute compaction task (#10205)
Currently, It always set 16 workers to execute the background compaction
jobs. The compaction task is an IO-and-CPU-bound job(load data, sort
block, calc ndv, bloomfilter...). In a resource-limited environment, it
will use up most of the hardware resources. Here we set the workers to
be max(CpuNumers/4, 1).

I will temporarily disable the memtable vector pool for this pr, because
the current memtable reuse has some problems, the actual observation
shows that the reuse rate is very low, and the relevant resources are
still pinned.