fix bug: delete same row multi times (#10534)
When multiple parent tables delete the same child table, it is possible to delete the same row of data multiple times. At present, when DN encounters this situation, it reports a w-w error.

This PR has been processed as follows:
when deleting the same table multiple times, it will point to a plan with a union, and then flow to the delete plan.

But it has not been processed yet:
When there are more than 2 children in union Node, I have created a new issue to track this case:
https://github.com/matrixorigin/matrixone/issues/10533

Approved by: @heni02, @nnsgmsone, @aunjgr