[refactor] clean up port configs.
For CN and DN, add new config port-base as the starting port for
the port number used by the service. Means that it is not neccesary
to config ports for ctl-service, lock-service and etc, port-base is
enough. But for compatibility, the new policy is enabled only when
port-base is configured. And if it not configured, the old configs
for the address work.

For logservice, we do not use the port-base for two reasons. First,
logservice will rarely have new services, and no new ports. Second,
it is not suiteable to use port-base in logservice as hakeeper-client
and gossip-seed-addresses needs to know the exact port numbers.

MO doc updated:

https://github.com/matrixorigin/matrixorigin.io.cn/pull/625