add default panic handler for log append task (#10988)
when ctrl+c shut down the log service, the dn service cannot append entries to the log service, after retrying 180s, the append task will panic.

however, all append tasks are submitted to a goroutine pool to be scheduled, and the pool will catch the panic through `recover`, and do some operations specified by users.

In the previous implementation, no panic handler is specified to the pool when calling the 'ants.NewPool, so the panic` can only crash the goroutine it belongs to instead of the dn service.

after the append task crashes, some other tasks will fall into the dead wait, which expected the append task to return and do some further work.

that is why the ctrl+c cannot shut down the dn service.

the fix is to specify a panic handler.

but it still needs to wait for 180s when the ctrl + c dn service or the log service cash in some situations.

Approved by: @XuPeng-SH