[feature] cn: add new config for init work state. (#11067)
Adding a CN instance and updating the label information of the CN
instance are not atomic operations, so there is a period of time
between when the CN is added to the hakeeper and when the label
information is updated that the label information is inaccurate, and
during this period of time it is necessary to make sure that the CN
can't be used incorrectly, and therefore it is necessary to set the
state of draining when the cn is added to the hakeeper to avoid
this problem. Therefore, we need to set the state to draining when
the cn is added to the hakeeper to avoid this problem.

Approved by: @zhangxu19830126, @reusee, @daviszhen