fix the sort key missing in tables having cluster by clause (#12019)
fix the sort key missing in tables having a `cluster by` clause.

the sort key only exists in memory and is constructed when creating schema by reading checkpoint data.

the sort key will be used later in `metadata_scan` and judging if it is needed to `mergesort` columns.

Approved by: @XuPeng-SH