[bug] incre: bootstrap fails as the wrong incre service instance. (#11861)
There is only one GlobalVariable for IncrementService even in
multi CN cluster which is in lauch mode. This causes one of the
CN instance get the wrong IncrementSerivce instanc which is created
by other CN instances.

So, we need to check if the IncrementService is the right one when
try to get it. If it is not, get the right one from GlobalVariable by service
ID, that is set when the instance is created.

Approved by: @qingxinhome, @daviszhen, @m-schen, @reusee, @XuPeng-SH, @zhangxu19830126, @nnsgmsone