fix txn leak in cloud env (#12028)
If a cn with the same uuid is restarted within a short period of time, it will lead to the possibility that the remote locks will not be released, because the heartbeat timeout of a remote lockservice cannot be detected. To solve this problem we use uuid+create-time as service id, then a cn reboot with the same uuid will also be considered as not a lockservice.

Approved by: @w-zr, @fengttt, @nnsgmsone, @reusee, @m-schen, @daviszhen