fix lockservice panic with try hold on empty lock (#12181)
Fix try hold on empty lock.  If a range lock has no holders, the all waiters need to be notified.  Because these waiters may be attempting to acquire a non-conflicting row locks or range locks between the current range.  Also, the range locks needs to be removed, otherwise an Empty Lock will be left in the lock table.

Approved by: @nnsgmsone