[bug] logtail: fix hang when logtail consumer reconnect. (#12401)
When logtail resumer occurs some kind of error, it will try to reconnect
to logtail server. Before reconnect, it aborts all running transactions.
Before rollback txns, txn client should cancel all timestamp waiters,
because there may be some transactions are trying to commit and
waiting for the new TS, but they cannot wait for any new TS as the
logtail consumer is in the status of reconnecting.

Approved by: @zhangxu19830126