[BugFix]: If SK=PK, don't use Varlena as column type (#12391)
This is a follow-up of https://github.com/matrixorigin/matrixone/pull/12365

Previously, when SK = PK, we created the `__mo_index_idx_col` column with the `T_varchar` type. But with this fix, we follow the same logic as the Unique Index Table. If len(SK)==1, then we use the **PK' col type** for `__mo_index_idx_col`

Approved by: @ouyuanning, @qingxinhome, @aressu1985, @m-schen