[refactor] logtail: some refactor logtail consumer and fix a bug (#12576)
1. refactor logtail consumer to fix hang problem: when restart TN
continuously，
which means CN has not connect to logtail server for the first time, TN
restart
again. In this case, logtail could not connect to logtail server
normally.
2. fix a bug that cause MO hang: when abort all txns, we should not
delete the active txn immediatly, which will cause this active txn will
not
be delete in callback function closeTxn(), in which, the number of user
txn
will be subtracted. When 4*CPU-num of active txns are not subtracted, MO
  is hang as the max-active-txns is triggered.