embedding the s3-writer returned object stats to the batch the CN will commit to TN. (#12780)
1. summary

in some cases, the CN will write S3 directly. after the writer is done, it has all the metadata of this object which can be reused later to avoid reloading object metadata from disk.

this PR enables the writer to construct an `object stats` when an object has been written down, and embedding these `stats` to the batch the CN will commit to TN.

and replaced part of the block info with this object stats in ranges.

Approved by: @m-schen, @nnsgmsone, @LeftHandCold, @XuPeng-SH