[enhancement] txnClient: it waits until txn client is in normal state. (#13138)
Add a new config "normal-state-no-wait" in CN to control whether it
returns an error or waits when the txn client is in pause state.

The configuration value is false by default, which means that if the txn
client is in pause state, it will wait until it changes to nromal state. Otherwise,
just return an error.

Approved by: @daviszhen, @zhangxu19830126, @reusee