support locate function (#13448)
The LOCATE() function in MySQL is used to find the position of a substring within a string. It returns the starting position of the first occurrence of the substring within the string. The syntax for the LOCATE() function is as follows:

LOCATE(substring, string, [start_position])

The function takes three parameters:
substring: This is the substring that you want to search for within the string.
string: This is the string within which you want to search for the substring.
start_position (optional): This is the position in the string where the search should start. If not specified, the search starts from the beginning of the string. The position is 1-based, meaning the first character has a position of 1.

Approved by: @m-schen, @heni02, @sukki37