enable pkfilter for auto increment type pk if none valus is null (#13967)
The reason why `cloud_device.real_time_position` still does not use pkfilter after changing the unique key to the primary key is that the primary key type is auto-increment.

There are constraints on the use of pkfilter for auto-increasing primary keys. When the inserted value is null or the insert statement does not have a corresponding column, mo need to wait for the preinsert operator to fill in the value. so pkfilter  can not be used.

But when the pk column of auto increment is inserted and the above constraints are satisfied, pkfilter can actually be used.

https://github.com/matrixorigin/MO-Cloud/issues/1329#issuecomment-1873285416

Approved by: @ouyuanning