Save a few lines with OperatorBase. (#14295)
It is a study of Go OOP.  We embed OperatorBase in
all Operators (Argument).   Data, method, etc that are
commen all operators goes here and only need to implement once.

It is about 1K lines of dup code.  Let's try to remove.
More important, later if we need to change anything of Base, we only need to do it once.

Approved by: @ouyuanning, @badboynt1, @m-schen, @nnsgmsone