GC Adapt Object (#14360)
This PR is because 1.1 added the object list and removed the block, but now the GC logic still follows the previous logic of consuming the block list. In order to make it work normally quickly and easily before GC reconstruction, the object is consumed as a block, which can be regarded as a design in which an object has only one block.

Approved by: @fengttt, @XuPeng-SH