a quick fix that avoids the failure of mo_table_size when the index table name is empty (#15404)
this is a quick fix for the issue of the indexTableName is empty.
the empty indexTableName causes the SQL parser err: table "" does not exist err and
then the mo_table_size call will fail.
this fix does not fix the issue but only avoids the failure caused by it.

cherry-pick to 1.1: https://github.com/matrixorigin/matrixone/pull/15403

Approved by: @m-schen