[fix] : remove pre check logic for create unique index for existed table (#15327)
When create a unique index for an existing table, there is no need to check it in advance, because it will also be checked when inserting data into the hidden table.

And the existing detection logic uses group by, which will fail when the amount of data is large.

Approved by: @nnsgmsone, @heni02, @m-schen