[Fix] : replace point locks with table lock in some cases (#15578)
sql like `insert into A select * from B`, will try to lock the value in `B`. When the number of these locks is too much, it will lead to bugs reflected in the issue. In the context of range locks being temporarily unavailable, use table lock to replace point lock

Previous pr's used the wrong stats, now it's standardized to do the judgement in the last step of building the logic plan.