[fix] error caused by parallel slicing in load data (#15786)
1. Rewrite the parallel split logic for CSV files to avoid exceptions caused by newline characters.
2. Adjust the configuration of csvParser to avoid Out of Memory; the current configuration has a buffer size of 1GB and reads 64MB of data per iteration.

The new partitioning logic can, under the best circumstances, split the file into the number of threads specified, and under the worst circumstances, it can only be split into a single piece (which is almost never the case).

Performance Testing:
The total volume of data is approximately 370GB. Performance tests were carried out on both the main branch and this branch.
Results for the main branch:
<img width="1085" alt="image-20240429164222669" src="https://github.com/matrixorigin/matrixone/assets/30954408/5b47970f-cd3e-4734-bd13-e01113ba6f90">
<img width="1080" alt="image-20240429171008150" src="https://github.com/matrixorigin/matrixone/assets/30954408/99755e35-081b-4951-b7f5-30786721d483">

Results for this branch:
<img width="1500" alt="image" src="https://github.com/matrixorigin/matrixone/assets/30954408/28e29724-0895-4a82-8e88-180b7f0caec5">
<img width="1514" alt="image" src="https://github.com/matrixorigin/matrixone/assets/30954408/3948a373-32ca-4ba7-8f1f-43105a6c150b">

Under normal circumstances, there is hardly any difference in performance between the two.

Approved by: @m-schen, @ouyuanning, @badboynt1, @aunjgr