fix Compile clear() method bug. (#17352)
the clear method of the Compile structure ,which will be called after query done, should not modify the internal properties of the proc, because this attribute is directly passed in from the outside. When a query fails and a retry occurs, this attribute will still be used.

Approved by: @ouyuanning, @sukki37