Fix GC MaxConsumed For Backup  (#17472)
During backup, if GC does not consume the most recent global checkpoint, the restored cluster will not be able to find the checkpoint to be used for GC and will not work. This is because backup only backs up data after the most recent global checkpoint. So raise the GC water level

Approved by: @XuPeng-SH, @daviszhen