Pipeline Refactor: table function (#18039)
This is a relatively large refactor.   Old table function impl is
broken.   Restructure the code.

## What type of PR is this?

- [ ] API-change
- [ ] BUG
- [ ] Improvement
- [ ] Documentation
- [ ] Feature
- [ ] Test and CI
- [x] Code Refactoring

## Which issue(s) this PR fixes:

issue #

## What this PR does / why we need it:

Refactor table function code. It is way worse than I expected, so I ends
up rewrite some code. I will be amazed if this thing works at all, but
good part is that this should not block real work in pipeline
refactoring. The code need to be redone, period.