[bug] proxy: session cannot be transferred after load data local infile (#18927)
In the case of "load data local infile" statement, client sends the first packet, then server sends response, which is "0xFB + filename", after that, client sends content of filename and an empty packet, at last, server sends OK packet. The sequence ID of this OK packet is not 1, and will cause the session cannot be transferred after this stmt finished. So, the solution is: when server sends 0xFB and the sequence ID of next packet is 3 bigger than last one, the next packet MUST be an OK packet, and the transfer is allowed.

Approved by: @zhangxu19830126