make txn do not wait in incrservice while the logtail is paused  (#19400)
Fix hung.  Consider the following timing:
1. logtail paused
2. cancel all active pipelines
3. reset logtail state to normal

When a pipeline in 2 is executing and a new transaction is opened in another goroutine to do something (e.g., update the cache of a auto-increment), the new transaction is blocked until the logtail state returns to normal. This creates a circular wait.

Approved by: @iamlinjunhong