[feature] logservice: implement archive logs. (#19460)
logservice implement archive logs feature.

if this feature is enabled by setting ArchiveLogEnabled to true,
the logs are archived to external storage before they are truncated.
And we can read the archived logs from the external storage even
if they are truncated.

Note that the logs are lost if the log replica is recreated.

Approved by: @zhangxu19830126