Remove bvt tag 18547 (#19648)
Clear a bvt tag.  We have normalized all string functions to max length MaxVarcharlen (64K).   Original test has a row
in the result that is longer than 64K, therefore the test, tag, and fix the result.

Original maxlen is never documented and we may need to extend it.   But if we want to extend it we should extend all
string functions in a consistent manner.   For now, it is 64K.

Approved by: @aressu1985