FullText Index:  Phrase search result do the exact match with the search string that keep the same words order as search string (#19871)
Bug fix: Phrase search should do exact match.  Word sequence such as "blue is not red" should not match with "blue is red" in phrase search.

To fix this:
for example, search the phrase "blue is red" with the content "blue is not red"
1. do the search same as before and all words "blue", "is", "red" found
2. keep the word order and position which is same as the search string to do the exact match.

"red" is not in same position.  position of "red" in search string is 8 and relative position of "red" in content is 12.

Approved by: @heni02, @sukki37