Plugin framework (#21141)
We need a plugin framework in SQL that allows to execute a custom plugin executable that can manipulate the data from (datalink or string) in database.  The result generated by the plugin can then be shown as a result in SQL.

For application,
1. generate embedding from LLM vector with source data (datalink or string) stored in database and then insert embedding into the table in database.
2. perform RAG application ask() function to LLM in SQL.  You can create a plugin executable to implement the ask function. The plugin will do the following:
a. Read the question from stdin
b. generate the embedding of the question from LLM
c. perform SQL with the LLM index from our database to get the N-best results
d. send N-best results and question to LLM to get the RAG answer
e. send the RAG answer to database
f. show the answer as result

Approved by: @aunjgr, @daviszhen, @qingxinhome, @zhangxu19830126, @badboynt1, @ouyuanning, @aressu1985, @heni02