skip to subscribe to the deleted table when calculating the mo table stats.  (#21252)
skip to subscribe to the deleted table when calculating the mo table stats.

for the deleted tables:
1. should not subscribe
2. even subscribed, should failed

However, in the current implementation, the mo table stats task may subscribe to a deleted table and succeed, which may cause some other bugs. So, it tries to skip to subscribe to a deleted table.

Approved by: @XuPeng-SH, @zhangxu19830126, @heni02