Support  'where tuple in tupleList' (#22249)
Before this PR, MO only supports `single colref in tuple`

After this PR, during the binding phase, `tuple in tupleList` will be rewritten
For example, `(a, b) in ((1, 2), (3, 4))` will be rewritten into `(a = 1 and b = 2) or (a = 3 and b = 4)`.
Implicit cast is also supported.

<img width="1236" height="274" alt="企业微信截图_17533448366588" src="https://github.com/user-attachments/assets/86bd3668-1eb9-4359-8d1c-84f4dd870dd5" />

Approved by: @heni02, @ouyuanning, @aunjgr, @XuPeng-SH